Fix testing database exists even if remote
The script 'database_exists' assumes that all databases are on the same
host by using the database configuration 'SYSTEM' to query the table
'pg_database'.  However, this cannot be assumed with the new database
configuration 'PROD_STANDBY' which specifically points to another host.

This patch makes the script 'database_exists' use the tested database
configuration to query the table 'information_schema_catalog_name'.